requirements.txt
beautifulsoup4
requests
pandas
html5lib
selenium
psycopg2
python-dotenv
flask
SQLAlchemy
Flask-SQLAlchemy
